refactor: moved quota tests to different file
